setup un tested flow

+++ database/migrations/2019_06_17_182135_create_cars_table.php
@@ -16,8 +16,8 @@ class CreateCarsTable extends Migration
         Schema::create('cars', function (Blueprint $table): void {
             $table->bigIncrements('id');
 
-            $table->unsignedInteger('company_id');
-            $table->unsignedInteger('model_id');
+            $table->unsignedBigInteger('company_id');
+            $table->unsignedBigInteger('model_id');
 
             $table->foreign('company_id')
                 ->references('id')
